+++ packages/noka/src/Application/index.ts
@@ -1,5 +1,3 @@
-/** @format */
-
 export * from "./Application";
 export * from "./ApplicationLike";
 export * from "./ApplicationOptions";