+++ packages/noka-tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # noka-tsconfig
 
+## 2.0.13
+
+### Patch Changes
+
+- 重构 config/logger/setuploader & 改进调试
+
 ## 2.0.12
 
 ### Patch Changes